requirements.txt
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
click==8.1.7
config==0.5.1
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.4
flwr==1.3.0
fonttools==4.42.1
fsspec==2023.9.1
grpcio==1.43.0
huggingface-hub==0.17.2
idna==3.4
importlib-resources==6.0.1
iterators==0.0.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
matplotlib==3.7.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
msgpack==1.0.5
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==23.1
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyparsing==3.1.1
python-configuration==0.8.2
python-dateutil==2.8.2
PyYAML==6.0.1
ray==1.11.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
safetensors==0.3.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
termcolor==2.3.0
timm==0.9.7
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tqdm==4.66.1
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
urllib3==2.0.5
yacs==0.1.8
zipp==3.16.2
